+++ apps/nextjs-website/src/components/organisms/WebinarDetailTemplate/WebinarDetailTemplate.tsx
@@ -18,6 +18,8 @@ import Typography from '@mui/material/Typography';
 import BlocksRendererClient from '@/components/molecules/BlocksRendererClient/BlocksRendererClient';
 import { pageToBreadcrumbs } from '@/helpers/breadcrumbs.helpers';
 import ProductBreadcrumbs from '@/components/atoms/ProductBreadcrumbs/ProductBreadcrumbs';
+import RelatedResources from '@/components/molecules/RelatedResources/RelatedResources';
+import QuestionsAndAnswers from '@/components/molecules/QuestionsAndAnswers/QuestionsAndAnswers';
 
 type WebinarDetailTemplateProps = {
   webinar: Webinar;
@@ -166,6 +168,35 @@ const WebinarDetailTemplate = ({ webinar }: WebinarDetailTemplateProps) => {
       {bodyContent}
       {speakerList}
       {startInfo}
+      {webinar.questionsAndAnswers && (
+        <QuestionsAndAnswers items={[...webinar.questionsAndAnswers]} />
+      )}
+      {webinar.relatedResources && (
+        <RelatedResources
+          title={webinar.relatedResources.title}
+          resources={webinar.relatedResources.resources.map((resource) => ({
+            title: resource.title,
+            description: {
+              title: resource.subtitle || '',
+              content: resource.description,
+            },
+            imagePath: resource.image?.url || '/images/hero.jpg',
+            link: {
+              label: resource.linkText,
+              href: resource.linkHref,
+            },
+            mobileImagePath: resource.image?.url || '/images/hero.jpg',
+          }))}
+          downloadableDocuments={(
+            webinar.relatedResources.downloadableDocuments || []
+          ).map((doc) => ({
+            title: doc.title,
+            size: doc.size,
+            downloadLink: doc.downloadLink,
+            tags: [{ label: doc.extension }],
+          }))}
+        />
+      )}
       {relatedLinks}
       <Snackbar
         open={!!error}