+++ apps/nextjs-website/src/components/templates/SolutionTemplate/SolutionTemplate.tsx
@@ -14,7 +14,8 @@ import FutureWebinarsShowcase from '@/components/organisms/FutureWebinarsShowcas
 import Stats from '@/components/atoms/Stats/Stats';
 import EContainer from '@/editorialComponents/EContainer/EContainer';
 import { BlocksContent } from '@strapi/blocks-react-renderer';
-import { Media } from '@/lib/types/media';
+import { Media } from '@/lib/strapi/codecs/MediaCodec';
+import NewsShowcase from '@/components/organisms/NewsShowcase/NewsShowcase';
 
 export type SolutionTemplateProps = {
   slug: string;
@@ -37,6 +38,19 @@ export type SolutionTemplateProps = {
   products: Pick<Product, 'logo' | 'slug' | 'name' | 'description'>[];
   webinars: Webinar[];
   bannerLinks: BannerLinkProps[];
+  successStories?: {
+    title: string;
+    subtitle?: string;
+    stories: {
+      title: string;
+      publishedAt?: Date;
+      path: string;
+      image?: {
+        url: string;
+        alternativeText?: string;
+      };
+    }[];
+  };
 };
 
 const SolutionTemplate = ({
@@ -50,6 +64,7 @@ const SolutionTemplate = ({
   products,
   webinars,
   bannerLinks,
+  successStories,
 }: SolutionTemplateProps) => {
   const { palette, spacing } = useTheme();
   const t = useTranslations();
@@ -92,12 +107,27 @@ const SolutionTemplate = ({
         />
       </EContainer>
       {bannerLinks && <BannerLinks bannerLinks={bannerLinks} />}
-      {stats && (
+      {stats.length > 0 && (
         <Stats
           maxWidth={265}
           items={stats.map((stat) => ({
             title: stat.title,
             description: stat.description,
+          }))}
+        />
+      )}
+      {successStories && (
+        <NewsShowcase
+          marginTop={8}
+          newsMarginTop={4}
+          title={successStories.title}
+          subtitle={successStories.subtitle}
+          items={successStories.stories.map((story) => ({
+            ...story,
+            link: {
+              url: story.path,
+              text: t('shared.readStory'),
+            },
           }))}
         />
       )}