+++ apps/nextjs-website/src/lib/strapi/codecs/CaseHistoriesComponentCodec.ts
@@ -1,0 +1,11 @@
+import * as t from 'io-ts';
+import { NullToUndefinedCodec } from '@/lib/strapi/codecs/NullToUndefinedCodec';
+import { CaseHistoryCodec } from '@/lib/strapi/caseHistoriesCodec';
+
+export const CaseHistoriesComponentCodec = t.strict({
+  title: t.string,
+  description: t.union([NullToUndefinedCodec, t.string]),
+  case_histories: t.strict({
+    data: t.array(CaseHistoryCodec),
+  }),
+});