+++ apps/nextjs-website/src/lib/types/media.ts
@@ -1,9 +1,0 @@
-export type Media = {
-  readonly name: string;
-  readonly alternativeText?: string;
-  readonly width: number;
-  readonly height: number;
-  readonly ext: string;
-  readonly mime: string;
-  readonly url: string;
-};