+++ apps/storybook-app/stories/templates/SolutionTemplate.stories.tsx
@@ -18,7 +18,8 @@ export default meta;
 
 const products = [
   {
-    name: 'PagoPA',
+    name: 'pagoPA',
+    shortName: 'pagoPa',
     description: 'Il portale per gli sviluppatori di PagoPA',
     slug: 'pagopa',
     logo: {
@@ -27,6 +28,7 @@ const products = [
   },
   {
     name: 'App IO',
+    shortName: 'App IO',
     description: 'Il portale per gli sviluppatori di App IO',
     slug: 'app-io',
     logo: {
@@ -35,6 +37,7 @@ const products = [
   },
   {
     name: 'SEND',
+    shortName: 'SEND',
     description: 'Il portale per gli sviluppatori di Send',
     slug: 'send',
     logo: {
@@ -122,7 +125,6 @@ export const Showcase: StoryObj<typeof SolutionTemplate> = {
           `<br />\n` +
           `<p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Fusce auctor enim vel sem fringilla, vitae malesuada nisi malesuada. Sed euismod augue id mauris aliquam, at dapibus lectus laoreet. Sed vel nulla vel risus gravida malesuada ac id tortor. Nulla facilisi. Vestibulum ante ipsum primis in faucibus orci luctus et ultrices posuere cubilia curae; Sed euismod, risus eget bibendum bibendum, quam nisi aliquam nisi, id congue lorem risus id nisi. Sed euismod, risus eget bibendum bibendum, quam nisi aliquam nisi, id congue lorem risus id nisi. Sed euismod, risus eget bibendum bibendum, quam nisi aliquam nisi, id congue lorem risus id nisi.</p>\n`,
         slug: 'test-2024-03-14',
-        isVisibleInHome: false,
         isVisibleInList: false,
         imagePath: 'https://dev.developer.pagopa.it/images/webinars.png',
         speakers: [