+++ apps/storybook-app/CHANGELOG.md
@@ -1,5 +1,27 @@
 # storybook-app
 
+## 0.7.0
+
+### Minor Changes
+
+- 53d26d1: Add QuestionsAndAnswers component
+- ed77bbf: Add RelatedResources and DownloadableDocuments components
+
+### Patch Changes
+
+- ec5f3f2: Fix issue with solution product
+
+## 0.6.0
+
+### Minor Changes
+
+- 82fe16d: Add Chat molecule component
+
+### Patch Changes
+
+- d6eff42: Quotes removed from tutorials model, fix issue with tutorial storybook
+- b3d75ad: Remove "page" word from template files
+
 ## 0.5.0
 
 ### Minor Changes